feat(article): 重构文章发布功能并优化关联数据处理
-重构了文章请求和响应结构，增加了更多文章元数据字段
- 优化了标签和分类的关联处理逻辑
- 更新了文章更新接口，使用 PUT 方法
- 添加了文件上传相关配置
-调整了前端界面和交互，以适应新的文章结构

+++ internal/models/article_category.go
@@ -3,8 +3,8 @@ package models
 // ArticleCategory 文章分类关联模型
 type ArticleCategory struct {
 	BaseModel
-	ArticleId  uint64 `json:"article_id" gorm:"not null;comment:文章id"`
-	CategoryId uint64 `json:"category_id" gorm:"not null;comment:分类id"`
+	ArticleId  int `json:"article_id" gorm:"not null;comment:文章id"`
+	CategoryId int `json:"category_id" gorm:"not null;comment:分类id"`
 }
 
 // TableName 指定表名
